updated js
